+++ Diglett/Diglett/HullGUI.cpp
@@ -19,20 +19,19 @@ void HullGUI::draw(
     if( position.y < 0 ) position.y = target.getSize().y + position.y;
     // Draw the background.
     sf::RectangleShape background = sf::RectangleShape( 
-            sf::Vector2f( GUI_SIZE.x, GUI_SIZE.y ) );
+            sf::Vector2f( (float) GUI_SIZE.x, (float) GUI_SIZE.y ) );
     background.setFillColor( BACKGROUND_COLOR );
-    background.setPosition( position.x, position.y );
+    background.setPosition( (float) position.x, (float) position.y );
     target.draw( background );
     // Fill according to current hull strength.
     float currentHullFraction = (float) player.getCurrentHull() / 
             (float) player.getMaxHull();
-    sf::RectangleShape fill = sf::RectangleShape( 
-            sf::Vector2f( 
-                currentHullFraction * ( GUI_SIZE.x - 2*BORDER_THICKNESS ), 
-                ( GUI_SIZE.y - 2*BORDER_THICKNESS ) ) );
+    sf::RectangleShape fill = sf::RectangleShape( sf::Vector2f( (float) 
+		    ( currentHullFraction * ( GUI_SIZE.x - 2*BORDER_THICKNESS ) ), 
+            (float) ( GUI_SIZE.y - 2*BORDER_THICKNESS ) ) );
     fill.setFillColor( FILL_COLOR );
-    fill.setPosition( position.x + BORDER_THICKNESS, 
-            position.y + BORDER_THICKNESS );
+    fill.setPosition( (float) ( position.x + BORDER_THICKNESS ), 
+            (float) ( position.y + BORDER_THICKNESS ) );
     target.draw( fill );
     // Draw the text.
     std::string textString = 
@@ -43,8 +42,8 @@ void HullGUI::draw(
     text.setStyle( sf::Text::Bold );
     text.setOrigin( text.getLocalBounds().width / 2, 
             text.getLocalBounds().height / 2 );
-    text.setPosition( position.x + (GUI_SIZE.x/2), 
-            position.y + (GUI_SIZE.y/2) );
+    text.setPosition( (float) ( position.x + (GUI_SIZE.x/2) ), 
+            (float) ( position.y + (GUI_SIZE.y/2) ) );
     target.draw( text );
 }
