+++ Diglett/Diglett/ShopGUI.cpp
@@ -17,23 +17,26 @@ ShopGUI::~ShopGUI(void) {
 void ShopGUI::draw( 
         sf::RenderTarget& target, 
         sf::RenderStates states ) const {
-    sf::Vector2i size = sf::Vector2i( 
-            ( 2 * BORDER + COLUMNS * ENTRY_SIZE.x ),
-            ( 2 * BORDER + ITEMS_PER_COL * ENTRY_SIZE.y ) );
+    sf::Vector2f size = sf::Vector2f( 
+            (float) ( 2 * BORDER + COLUMNS * ENTRY_SIZE.x ),
+            (float) ( 2 * BORDER + ITEMS_PER_COL * ENTRY_SIZE.y ) );
 
-    sf::Vector2i position = sf::Vector2i(
+    sf::Vector2f position = sf::Vector2f(
             ( target.getSize().x / 2 ) - size.x / 2,
             ( target.getSize().y / 2 ) - size.y / 2 );
     std::vector<Item *> contents = inventory.getContents();
     // Draw the inventory background.
-    background.setPosition( sf::Vector2f( position.x, position.y - TOP_BORDER ) );
+    background.setPosition( 
+		    sf::Vector2f( position.x, position.y - (float) TOP_BORDER ) );
     target.draw( background );
     for( int i = 0; i < inventory.getCurrentSize(); i++ ) {
         // Calculate the position of the sprite.
-        int spritePositionX = 
-            position.x + BORDER + ( i / ITEMS_PER_COL ) * ENTRY_SIZE.x;
-        int spritePositionY =
-            position.y + BORDER + ( i % ITEMS_PER_COL ) * ENTRY_SIZE.y;
+        float spritePositionX = 
+                position.x + 
+				(float) ( BORDER + ( i / ITEMS_PER_COL ) * ENTRY_SIZE.x );
+        float spritePositionY =
+                position.y + 
+				(float) ( BORDER + ( i % ITEMS_PER_COL ) * ENTRY_SIZE.y );
         // Draw the sprite.
         contents[i]->getSprite()->setPosition( 
                 spritePositionX, spritePositionY );
@@ -71,7 +74,7 @@ const int ShopGUI::TOP_BORDER = 40;
 const int ShopGUI::BORDER = 10;
 const int ShopGUI::COLUMNS = 4;
 const int ShopGUI::ITEMS_PER_COL = 8;
-const sf::Vector2i ShopGUI::ENTRY_SIZE = sf::Vector2i( 128, 19 );
+const sf::Vector2f ShopGUI::ENTRY_SIZE = sf::Vector2f( 128, 19 );
 const int ShopGUI::SPRITE_SEPARATION = 6;
 const int ShopGUI::TEXT_SIZE = 12;
 const sf::Color ShopGUI::TEXT_COLOR = sf::Color( 255, 255, 255, 255 );
